move ClientMemberListProvider
